[ci skip] Refactor, continued
