+++ server/queues/processors/backlinks.js
@@ -1,9 +1,9 @@
 // @flow
-import { Document, Backlink } from "../models";
-import { Op } from "../sequelize";
-import type { DocumentEvent, RevisionEvent } from "../types";
-import parseDocumentIds from "../utils/parseDocumentIds";
-import slugify from "../utils/slugify";
+import { Document, Backlink } from "../../models";
+import { Op } from "../../sequelize";
+import type { DocumentEvent, RevisionEvent } from "../../types";
+import parseDocumentIds from "../../utils/parseDocumentIds";
+import slugify from "../../utils/slugify";
 
 export default class Backlinks {
   async on(event: DocumentEvent | RevisionEvent) {
