+++ server/queues/processors/revisions.js
@@ -1,8 +1,8 @@
 // @flow
 import invariant from "invariant";
-import revisionCreator from "../commands/revisionCreator";
-import { Revision, Document, User } from "../models";
-import type { DocumentEvent, RevisionEvent } from "../types";
+import revisionCreator from "../../commands/revisionCreator";
+import { Revision, Document, User } from "../../models";
+import type { DocumentEvent, RevisionEvent } from "../../types";
 
 export default class Revisions {
   async on(event: DocumentEvent | RevisionEvent) {
