+++ server/services/index.js
@@ -1,0 +1,6 @@
+// @flow
+import web from "./web";
+import websockets from "./websockets";
+import worker from "./worker";
+
+export default { web, websockets, worker };
